Detect changeset at point
Instead of taking the changeset from the beginning of a `sqitch.plan`
line, use the `thingatpt` library to detect the changeset "at point".
This allows us to jump to files from within the dependency list on a
changeset line.

The current implementation isn't perfect because it doesn't recognize
version-tagged changeset labels. For instance, "foo@0.0.1" with point
somewhere in "foo" only gets recognized as "foo". Being able to deal
with dependency lists is nice, though, and seems like a decent tradeoff
for now.